+++ src/aux-vm/managers/RecordsManager.ts
@@ -86,6 +86,7 @@ import { sha256 } from 'hash.js';
 import stringify from '@casual-simulation/fast-json-stable-stringify';
 import '@casual-simulation/aux-common/BlobPolyfill';
 import type { Observable } from 'rxjs';
+import { mergeWith } from 'rxjs';
 import {
     ReplaySubject,
     Subject,
@@ -2694,11 +2695,19 @@ export class RecordsManager {
                 client.connectionState.pipe(filter((c) => c.connected))
             );
 
+            const disconnected = client.connectionState.pipe(
+                filter((c) => !c.connected),
+                map(() => {
+                    throw new Error('The request encountered an error.');
+                })
+            );
+
             const id = this._httpRequestId++;
             const promise = firstValueFrom(
-                client
-                    .event('http_response')
-                    .pipe(filter((response) => response.id === id))
+                client.event('http_response').pipe(
+                    mergeWith(disconnected),
+                    filter((response) => response.id === id)
+                )
             );
 
             client.send({
@@ -2747,7 +2756,16 @@ export class RecordsManager {
         );
 
         const id = this._httpRequestId++;
+
+        const disconnected = client.connectionState.pipe(
+            filter((c) => !c.connected),
+            map(() => {
+                throw new Error('The request encountered an error.');
+            })
+        );
+
         const responses = client.event('http_partial_response').pipe(
+            mergeWith(disconnected),
             filter((response) => response.id === id),
             takeWhile((response) => !response.final),
             map((response) => {